Update the ConfigMap Zone Sync documentation
This commit updates the ConfigMap Zone Sync documentation by changing
the order and formatting of some information. Previously, all of the
content was in a note call-out after the table of keys and descriptions.

Now, the important information about TLS availability is part of the
main text, and the note call-out for a possible error is instead a
caution call-out just before the table.

Previously, there was a somewhat redundant (And broken) link. The
frontmatter data of the page has also been updated.